Move animation builder to separate dir (#83)
